[F] - Fix totalTimeListened values

+++ src/Database/UltimatePlaylist.Common.Infrastructure/Views/GeneralSongDataProcedureView.cs
@@ -26,7 +26,7 @@
 
         public int UniquePlays { get; set; }
 
-        public long? TotalTimeListened { get; set; }
+        public int? TotalTimeListened { get; set; }
 
         public string CoverUrl { get; set; }
     }
